Split up test helper into individual tests
